creating Semantic actions plugin  BDRE-387

+++ plugin-manager/src/main/java/com/wipro/ats/bdre/pm/FSOperations.java
@@ -4,12 +4,9 @@ import com.wipro.ats.bdre.md.pm.beans.FS;
 import org.apache.log4j.Logger;
 
 import java.io.File;
-import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.FileSystems;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 
 /**
@@ -17,16 +14,24 @@ import java.nio.file.StandardCopyOption;
  */
 public class FSOperations {
     private static final Logger LOGGER = Logger.getLogger(PluginManagerMain.class);
-    public void copyAction(FS fs,String pluginDescriptorJSON){
+    public void copyAction(FS fs,String pluginDescriptorJSON) throws FileNotFoundException{
 
         try{
             String homeDir = System.getProperty("user.home");
             File destFile = new File(homeDir + "/" + fs.getDestinationLocation());
+            File destDir = new File((homeDir + "/" + fs.getDestinationLocation()).substring(0,(homeDir + "/" + fs.getDestinationLocation()).lastIndexOf("/")));
+            destDir.mkdirs();
             File srcFile = new File(pluginDescriptorJSON + "/" + fs.getSourceLocation());
-            boolean fileCreated = destFile.createNewFile();
-            Files.copy(srcFile.toPath(),destFile.toPath(),StandardCopyOption.REPLACE_EXISTING);
+            if (srcFile.exists()) {
+                boolean fileCreated = destFile.createNewFile();
+                Files.copy(srcFile.toPath(), destFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            } else {
+                throw new FileNotFoundException("file not found");
+            }
+
         }catch (IOException f){
             LOGGER.error("file not found" + f);
+            throw new FileNotFoundException("file not found");
         }
     }
